style(green): moved tests to their own folder for easier cloc checks
